# public/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "description",
        "version": "0.5.0",
        "title": "China Cipher Messenger Server",
        "contact": {
            "email": "t105590035@ntut.org.tw"
        },
        "license": {
            "name": "MIT License",
            "url": "https://github.com/ChinaCipher/messenger-server/blob/master/LICENSE"
        }
    },
    "host": "ccm.ntut.com.tw",
    "basePath": "/api",
...
    "schemes": [
        "https"
    ],
    "paths": {
...
        "/session": {
            "get": {
                "tags": [
                    "UserApi"
                ],
                "summary": "取得會話驗證碼",
                "description": "長度為 22 個 ASCII 字元",
                "operationId": "get_session_code",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Session.Info"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "UserApi"
                ],
                "summary": "建立會話",
                "description": "`password` 格式為：`bcrypt(bcrypt(明文密碼, sha256(用戶名稱).slice(0, 22)), 會話驗證碼)`",
                "operationId": "login",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "所需的資料",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Login.Request"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/User.Profile.Private"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Test.Message"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "UserApi"
                ],
                "summary": "移除會話",
                "description": "",
                "operationId": "logout",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Test.Message"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Test.Message": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "訊息內容",
                    "type": "string"
                }
            }
        },
        "Session.Info": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "會話驗證碼",
                    "type": "string"
                },
                "profile": {
                    "description": "使用者資訊",
                    "$ref": "#/definitions/User.Profile.Private"
                }
            }
        },
...
        "Login.Request": {
            "type": "object",
            "properties": {
                "username": {
                    "description": "使用者名稱",
                    "type": "string"
                },
                "password": {
                    "description": "使用者密文密碼",
                    "type": "string"
                }
            }
        },
        "User.Profile": {
            "type": "object",
            "properties": {
                "avatar": {
                    "description": "大頭貼網址",
                    "type": "string"
                },
                "username": {
                    "description": "使用者名稱",
                    "type": "string"
                },
                "nickname": {
                    "description": "使用者暱稱",
                    "type": "string"
                }
            }
        },
...
        "User.Profile.Private": {
            "type": "object",
            "properties": {
                "user": {
                    "description": "使用者資訊",
                    "$ref": "#/definitions/User.Profile"
                },
                "privateKey": {
                    "description": "加密過後的使用者私鑰",
                    "type": "string"
                }
            }
        },
...
    }
}